Implemented interval to repeatedly fetch the prices for a currency
This feature automatically begins an interval whe a user selects a
currency.  The prices are fetched every minute so we keep the current
price and chart up to date.  Users will visually see the price, chart
and percentage increase/decrease change, keeping them up to date.

- Created a class method that can be bound to the class, keeping the
variables bound to it.
- Updated the startTicker method to bind the new class method for
fetching prices and pass it into the interval function.
- Created a method that will clear any live interval, fetch the prices
and then start the interval.  Updated the currency select's onChange event with a call to this method.